KE-15688 add argument --after-truncate for metadata restore

+++ src/tool/src/main/java/io/kyligence/kap/tool/MetadataTool.java
@@ -42,9 +42,6 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import io.kyligence.kap.common.persistence.metadata.AuditLogStore;
-import io.kyligence.kap.common.util.AddressUtil;
-import io.kyligence.kap.tool.util.ScreenPrintUtil;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.OptionGroup;
 import org.apache.commons.cli.Options;
@@ -74,10 +71,13 @@ import io.kyligence.kap.common.metrics.NMetricsCategory;
 import io.kyligence.kap.common.metrics.NMetricsGroup;
 import io.kyligence.kap.common.metrics.NMetricsName;
 import io.kyligence.kap.common.persistence.ImageDesc;
+import io.kyligence.kap.common.persistence.metadata.AuditLogStore;
 import io.kyligence.kap.common.persistence.transaction.UnitOfWork;
 import io.kyligence.kap.common.persistence.transaction.UnitOfWorkParams;
+import io.kyligence.kap.common.util.AddressUtil;
 import io.kyligence.kap.common.util.MetadataChecker;
 import io.kyligence.kap.metadata.project.UnitOfAllWorks;
+import io.kyligence.kap.tool.util.ScreenPrintUtil;
 import lombok.Getter;
 import lombok.val;
 import lombok.var;
@@ -100,6 +100,10 @@ public class MetadataTool extends ExecutableApplication {
 
     private static final Option OPERATE_RESTORE = OptionBuilder.getInstance()
             .withDescription("Restore metadata from local path or HDFS path").isRequired(false).create("restore");
+
+    private static final Option OPTION_AFTER_TRUNCATE = OptionBuilder.getInstance()
+            .withDescription("Restore overwrite metadata from local path or HDFS path (optional)").isRequired(false)
+            .withLongOpt("after-truncate").hasArg(false).create("d");
 
     private static final Option OPTION_DIR = OptionBuilder.getInstance().hasArg().withArgName("DIRECTORY_PATH")
             .withDescription("Specify the target directory for backup and restore").isRequired(false).create("dir");
@@ -147,6 +151,7 @@ public class MetadataTool extends ExecutableApplication {
         options.addOption(FOLDER_NAME);
         options.addOption(OPERATE_COMPRESS);
         options.addOption(OPTION_EXCLUDE_TABLE_EXD);
+        options.addOption(OPTION_AFTER_TRUNCATE);
     }
 
     public static void backup(KylinConfig kylinConfig) throws IOException {
@@ -165,7 +170,7 @@ public class MetadataTool extends ExecutableApplication {
 
     public static void restore(KylinConfig kylinConfig, String folder) throws IOException {
         val tool = new MetadataTool(kylinConfig);
-        tool.execute(new String[] { "-restore", "-dir", folder });
+        tool.execute(new String[] { "-restore", "-dir", folder, "--after-truncate" });
     }
 
     public static void main(String[] args) throws Exception {
@@ -203,7 +208,8 @@ public class MetadataTool extends ExecutableApplication {
                 backup(optionsHelper);
             } catch (Exception be) {
                 if (isGlobal) {
-                    NMetricsGroup.hostTagCounterInc(NMetricsName.METADATA_BACKUP_FAILED, NMetricsCategory.GLOBAL, GLOBAL);
+                    NMetricsGroup.hostTagCounterInc(NMetricsName.METADATA_BACKUP_FAILED, NMetricsCategory.GLOBAL,
+                            GLOBAL);
                 } else {
                     NMetricsGroup.hostTagCounterInc(NMetricsName.METADATA_BACKUP_FAILED, NMetricsCategory.PROJECT,
                             optionsHelper.getOptionValue(OPTION_PROJECT));
@@ -212,8 +218,8 @@ public class MetadataTool extends ExecutableApplication {
             } finally {
                 if (isGlobal) {
                     NMetricsGroup.hostTagCounterInc(NMetricsName.METADATA_BACKUP, NMetricsCategory.GLOBAL, GLOBAL);
-                    NMetricsGroup.hostTagCounterInc(NMetricsName.METADATA_BACKUP_DURATION, NMetricsCategory.GLOBAL, GLOBAL,
-                            System.currentTimeMillis() - startAt);
+                    NMetricsGroup.hostTagCounterInc(NMetricsName.METADATA_BACKUP_DURATION, NMetricsCategory.GLOBAL,
+                            GLOBAL, System.currentTimeMillis() - startAt);
                 } else {
                     NMetricsGroup.hostTagCounterInc(NMetricsName.METADATA_BACKUP, NMetricsCategory.PROJECT,
                             optionsHelper.getOptionValue(OPTION_PROJECT));
@@ -223,7 +229,7 @@ public class MetadataTool extends ExecutableApplication {
             }
 
         } else if (optionsHelper.hasOption(OPERATE_RESTORE)) {
-            restore(optionsHelper);
+            restore(optionsHelper, optionsHelper.hasOption(OPTION_AFTER_TRUNCATE));
         } else {
             throw new KylinException(INVALID_SHELL_PARAMETER, "The input parameters are wrong");
         }
@@ -311,7 +317,7 @@ public class MetadataTool extends ExecutableApplication {
         backupMetadataStore.dump(backupResourceStore);
         logger.info("backup successfully at {}", backupPath);
     }
-    
+
     private long getOffset(AuditLogStore auditLogStore) {
         long offset = 0;
         if (kylinConfig.isUTEnv())
@@ -320,7 +326,7 @@ public class MetadataTool extends ExecutableApplication {
             offset = auditLogStore.getStartId() == 0 ? resourceStore.getOffset() : auditLogStore.getStartId();
         return offset;
     }
-    
+
     private void backupProjects(NavigableSet<String> projectFolders, ResourceStore backupResourceStore,
             boolean excludeTableExd) throws InterruptedException {
         for (String projectPath : projectFolders) {
@@ -375,7 +381,8 @@ public class MetadataTool extends ExecutableApplication {
         }
     }
 
-    private void restore(OptionsHelper optionsHelper) throws IOException {
+    private void restore(OptionsHelper optionsHelper, boolean delete) throws IOException {
+        logger.info("Restore metadata with delete : {}", delete);
         val project = optionsHelper.getOptionValue(OPTION_PROJECT);
         val restorePath = optionsHelper.getOptionValue(OPTION_DIR);
 
@@ -392,12 +399,13 @@ public class MetadataTool extends ExecutableApplication {
         if (!verifyResult.isQualified()) {
             throw new RuntimeException(verifyResult.getResultMessage() + "\n the metadata dir is not qualified");
         }
-        restore(resourceStore, restoreResourceStore, project);
+        restore(resourceStore, restoreResourceStore, project, delete);
         backup(kylinConfig);
 
     }
 
-    public static void restore(ResourceStore currentResourceStore, ResourceStore restoreResourceStore, String project) {
+    public static void restore(ResourceStore currentResourceStore, ResourceStore restoreResourceStore, String project,
+            boolean delete) {
         if (StringUtils.isBlank(project)) {
             logger.info("start to restore all projects");
             var srcProjectFolders = restoreResourceStore.listResources("/");
@@ -414,9 +422,8 @@ public class MetadataTool extends ExecutableApplication {
                 val projectName = Paths.get(projectPath).getName(0).toString();
                 val destResources = currentResourceStore.listResourcesRecursively(projectPath);
                 val srcResources = restoreResourceStore.listResourcesRecursively(projectPath);
-                UnitOfWork.doInTransactionWithRetry(
-                        () -> doRestore(currentResourceStore, restoreResourceStore, destResources, srcResources),
-                        projectName, 1);
+                UnitOfWork.doInTransactionWithRetry(() -> doRestore(currentResourceStore, restoreResourceStore,
+                        destResources, srcResources, delete), projectName, 1);
             }
 
         } else {
@@ -437,22 +444,22 @@ public class MetadataTool extends ExecutableApplication {
             Set<String> finalGlobalDestResources = globalDestResources;
 
             UnitOfWork.doInTransactionWithRetry(() -> doRestore(currentResourceStore, restoreResourceStore,
-                    finalGlobalDestResources, globalSrcResources), UnitOfWork.GLOBAL_UNIT, 1);
+                    finalGlobalDestResources, globalSrcResources, delete), UnitOfWork.GLOBAL_UNIT, 1);
 
             val projectPath = "/" + project;
             val destResources = currentResourceStore.listResourcesRecursively(projectPath);
             val srcResources = restoreResourceStore.listResourcesRecursively(projectPath);
 
             UnitOfWork.doInTransactionWithRetry(
-                    () -> doRestore(currentResourceStore, restoreResourceStore, destResources, srcResources), project,
-                    1);
+                    () -> doRestore(currentResourceStore, restoreResourceStore, destResources, srcResources, delete),
+                    project, 1);
         }
 
         logger.info("restore successfully");
     }
 
     private static int doRestore(ResourceStore currentResourceStore, ResourceStore restoreResourceStore,
-            Set<String> destResources, Set<String> srcResources) throws IOException {
+            Set<String> destResources, Set<String> srcResources, boolean delete) throws IOException {
         val threadViewRS = ResourceStore.getKylinMetaStore(KylinConfig.getInstanceFromEnv());
 
         //check destResources and srcResources are null,because  Sets.difference(srcResources, destResources) will report NullPointerException
@@ -471,10 +478,11 @@ public class MetadataTool extends ExecutableApplication {
             val metadataRaw = restoreResourceStore.getResource(res);
             threadViewRS.checkAndPutResource(res, metadataRaw.getByteSource(), raw.getMvcc());
         }
-
-        val deleteRes = Sets.difference(destResources, srcResources);
-        for (val res : deleteRes) {
-            threadViewRS.deleteResource(res);
+        if (delete) {
+            val deleteRes = Sets.difference(destResources, srcResources);
+            for (val res : deleteRes) {
+                threadViewRS.deleteResource(res);
+            }
         }
 
         return 0;
